apps/chapter1/serializers.py
from abc import ABC, ABCMeta

from django.contrib import admin

from .models import Student

from rest_framework import serializers


class StudentSerializer(serializers.Serializer):
    id = serializers.IntegerField(required=False)
    name = serializers.CharField(max_length=100)
    roll = serializers.IntegerField()
    city = serializers.CharField(max_length=100)

    def create(self, validated_data):
        return Student.objects.create(**validated_data)



