pyproject.toml
[tool.poetry]
name = "lesson1"
version = "0.1.0"
description = "A practise project for django"
authors = ["Mithun Adhikari"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
